#10077: Overload binary lt in unary lt
