--- docs/source/ttnn/ttnn/ttnn/binary_lt_bw.rst
@@ -1,6 +1,0 @@
-.. _ttnn.binary_lt_bw:
-
-ttnn.binary_lt_bw
-#################
-
-.. autofunction:: ttnn.binary_lt_bw
