+++ ttnn/cpp/ttnn/operations/eltwise/binary_backward/device/binary_backward_op.hpp
@@ -33,7 +33,7 @@ enum class BinaryBackwardOpType {
     RSUB_BW,
     BIAS_GELU_BW,
     GT_BW,
-    BINARY_LT_BW,
+    LT_BW,
     BINARY_NE_BW,
     BINARY_GE_BW,
     MIN_BW,
